Revert "refactor: 수정 요청 DTO 이름 정리"
This reverts commit 46fcbbbee24ad299fb3990cf734f408025edaf6b.
