+++ src/main/java/gg/agit/konect/domain/club/controller/ClubRegistrationRequestApi.java
@@ -3,7 +3,7 @@ package gg.agit.konect.domain.club.controller;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 
-import gg.agit.konect.domain.club.dto.ClubInformationUpdateRequest;
+import gg.agit.konect.domain.club.dto.ClubInformationUpdateRequestDto;
 import gg.agit.konect.domain.club.dto.ClubRegistrationRequestDto;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 
@@ -15,6 +15,6 @@ public interface ClubRegistrationRequestApi {
 
     ResponseEntity<Void> requestClubInformationUpdate(
         @PathVariable(name = "clubId") Integer clubId,
-        @RequestBody ClubInformationUpdateRequest request
+        @RequestBody ClubInformationUpdateRequestDto request
     );
 }
